+++ src/shared/validators/auth.ts
@@ -1,4 +1,3 @@
-// src/shared/validators/auth.ts
 import { z } from "zod";
 
 export const emailSchema = z
